+++ spec/lib/stellar/client_spec.rb
@@ -11,6 +11,14 @@ describe Stellar::Client do
       expect(headers["Accept"]).
         to eq "application/hal+json,application/problem+json,application/json"
     end
+
+    it "has 'X-Client-Name'" do
+      expect(headers["X-Client-Name"]).to eq "ruby-stellar-sdk"
+    end
+
+    it "has 'X-Client-Version'" do
+      expect(headers["X-Client-Version"]).to eq Stellar::VERSION
+    end
   end
 
   describe "#default_testnet" do